requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart

motor
pydantic

black
isort